+++ gocd_cli/commands/pipeline/__init__.py
@@ -1,3 +1,7 @@
+from __future__ import print_function
+
+import time
+
 from gocd_cli.command import BaseCommand
 from gocd_cli.utils import get_settings
 
@@ -34,22 +38,51 @@ class Trigger(BaseCommand):
           be passed to the pipeline.
           Example: SHARED_SECRET=Mellon,SECOND=Really Im first
         secure_variables: A comma separated list of key=value pairs.
+        wait_until_finished: Will wait until the pipeline finishes running and
+          then exit 0 on success, and 2 on failure. The console.log will
+          be output for each stage in order.
+        verbose: Will print a . every tick when wait_until_finished is true
     """
     usage_summary = 'Triggers the named pipeline'
 
-    def __init__(self, server, name, unlock=False, variables=None, secure_variables=None):
+    _tick = 30  # seconds
+
+    def __init__(self, server, name, unlock=False, variables=None, secure_variables=None,
+                 wait_until_finished=False, verbose=False):
         self.pipeline = server.pipeline(name)
         self.unlock = str(unlock).lower().strip() == 'true'
         self.variables = self._convert_to_dict(variables)
         self.secure_variables = self._convert_to_dict(secure_variables)
+        self.wait_until_finished = str(wait_until_finished).lower().strip() == 'true'
+        self.verbose = str(verbose).lower().strip() == 'true'
 
     def run(self):
         if self.unlock:
             unlock_pipeline(self.pipeline)
 
-        return self.pipeline.schedule(
+        response = self.pipeline.schedule(
             variables=self.variables,
-            secure_variables=self.secure_variables
+            secure_variables=self.secure_variables,
+            return_new_instance=self.wait_until_finished,
+        )
+
+        if not self.wait_until_finished and response.is_ok:
+            return self._return_value('', exit_code=0)
+        elif not response.is_ok:
+            return self._return_value(response.body.strip(), response.is_ok)
+
+        instance_id = response['counter']
+        while not self._stages_finished(response):
+            if self.verbose:
+                print('.', end='')
+            time.sleep(self._tick)
+            response = self.pipeline.instance(instance_id)
+
+        self._print_job_output(response)
+
+        return self._return_value(
+            False,
+            self._run_successful(response),
         )
 
     def _convert_to_dict(self, args):
@@ -66,6 +99,26 @@ class Trigger(BaseCommand):
             variables[k] = v
 
         return variables
+
+    def _stages_finished(self, response):
+        for stage in response['stages']:
+            if stage['result'] not in self.pipeline.final_results:
+                return False
+
+        return True
+
+    def _run_successful(self, response):
+        for stage in response['stages']:
+            if stage['result'] == 'Failed':
+                return False
+
+        return True
+
+    def _print_job_output(self, instance):
+        for metadata, output in self.pipeline.console_output(instance):
+            job_masthead = ', '.join(('{0}="{1}"'.format(k, v) for k, v in metadata.items()))
+            print('\n\n=== {0} ===\n\n'.format(job_masthead))
+            print(output.strip())
 
 
 class Unlock(BaseCommand):
@@ -175,10 +228,10 @@ class List(BaseCommand):
         for pipeline in self.server.pipeline_groups().pipelines:
             status = self.server.pipeline(pipeline).status()
             if not status:
-                print 'Error getting status for "{0}"'.format(pipeline)
+                print('Error getting status for "{0}"'.format(pipeline))
                 exit(3)
 
-            print '{0}: {1}'.format(pipeline, self._format_status(status.payload))
+            print('{0}: {1}'.format(pipeline, self._format_status(status.payload)))
 
     def _format_status(self, status):
         return ', '.join(('{0}={1}'.format(k, v) for k, v in status.items()))