+++ CHANGELOG.rst
@@ -2,6 +2,21 @@
 Change Log
 ==========
 
+`0.10.0`_ - 2015-11-xx
+======================
+
+**Added**
+
+* Add ``wait-until-finished`` to pipeline trigger, which waits around
+  for the pipeline run to finalize and then outputs ``console.log`` from
+  each stage/job. Will exit 0 on success and 2 on failure.
+
+  Usage:
+
+  .. code-block:: shell
+
+      $ gocd pipeline trigger Simple-Pipeline --wait_until_finished=true
+
 `0.9.1`_ - 2015-11-06
 =====================
 
@@ -84,6 +99,7 @@ Changed
 
 Nothing much to say here, initial public release. :)
 
+.. _`0.10.0`: https://github.com/gaqzi/gocd-cli/compare/v0.9.1...v0.10.0
 .. _`0.9.1`: https://github.com/gaqzi/gocd-cli/compare/v0.9.0...v0.9.1
 .. _`0.9.0`: https://github.com/gaqzi/gocd-cli/compare/v0.8.0...v0.9.0
 .. _`0.8.0`: https://github.com/gaqzi/gocd-cli/compare/v0.7.1...v0.8.0